Core/Buildsystem: Add suport for vs2015 Core/Dep: remove mysqllite dependency, now user MUST install mysql server to use included libs Also remove lots of warnings in vs2015

+++ src/server/game/AI/SmartScripts/SmartScript.cpp
@@ -534,7 +534,7 @@ void SmartScript::ProcessAction(SmartScriptHolder& e, Unit* unit, uint32 var0, u
 						if (me->GetDistance((*itr)->ToUnit()) > spellInfo->GetMaxRange(true) ||
 							me->GetDistance((*itr)->ToUnit()) < spellInfo->GetMinRange(true) ||
 							!me->ToUnit()->IsWithinLOSInMap((*itr)->ToUnit()) ||
-							mana < spellInfo->CalcPowerCost(me, spellInfo->GetSchoolMask(), me->GetSpellPowerEntryBySpell(spellInfo)))
+							mana < (int32)spellInfo->CalcPowerCost(me, spellInfo->GetSchoolMask(), me->GetSpellPowerEntryBySpell(spellInfo)))
 							_allowMove = true;
 
 						CAST_AI(SmartAI, me->AI())->SetCombatMove(_allowMove);
